+++ src/003_request-dom-events.js
@@ -16,13 +16,20 @@ function bindRequestEvents(container, options) {
 
 	var handler;
 	if (options.links !== false) {
+		// anchor-click handler
 		handler = { name: 'click', handleEvent: Local__clickHandler, container: container };
 		container.addEventListener('click', handler, false);
 		container.__localEventHandlers.push(handler);
 	}
 	if (options.forms !== false) {
+		// submitter tracking
+		handler = { name: 'click', handleEvent: Local__submitterTracker, container: container };
+		container.addEventListener('click', handler, true); // must be on capture to happen in time
+		container.__localEventHandlers.push(handler);
+		// submit handler
 		handler = { name: 'submit', handleEvent: Local__submitHandler, container: container };
 		container.addEventListener('submit', handler, false);
+		container.__localEventHandlers.push(handler);
 	}
 }
 
@@ -43,7 +50,6 @@ function unbindRequestEvents(container) {
 // transforms click events into request events
 function Local__clickHandler(e) {
 	if (e.button !== 0) { return; } // handle left-click only
-	local.util.trackFormSubmitter(e.target);
 	var request = local.util.extractRequest.fromAnchor(e.target);
 	if (request && ['_top','_blank'].indexOf(request.target) !== -1) { return; }
 	if (request) {
@@ -52,6 +58,13 @@ function Local__clickHandler(e) {
 		local.util.dispatchRequestEvent(e.target, request);
 		return false;
 	}
+}
+
+// INTERNAL
+// marks the submitting element (on click capture-phase) so the submit handler knows who triggered it
+function Local__submitterTracker(e) {
+	if (e.button !== 0) { return; } // handle left-click only
+	local.util.trackFormSubmitter(e.target);
 }
 
 // INTERNAL